deployment/requirements.txt
streamlit
pandas
numpy
scikit-learn
imbalanced-learn
daal4py
feature_engine
tensorflow-cpu